Move AddTodoForm to containers directory
